refactor: code snippet functions to util file
