+++ apps/web/src/pages/templates/components/TriggerSnippetTabs.tsx
@@ -5,8 +5,8 @@ import * as get from 'lodash.get';
 
 import { INotificationTrigger, INotificationTriggerVariable, TemplateVariableTypeEnum } from '@novu/shared';
 
-import { API_ROOT } from '../../../config';
 import { colors, Tabs } from '@novu/design-system';
+import { createCurlSnippet, createNodeSnippet } from '../../../utils/codeSnippets';
 
 const NODE_JS = 'Node.js';
 const CURL = 'Curl';
@@ -44,33 +44,6 @@ export function TriggerSnippetTabs({ trigger }: { trigger: INotificationTrigger 
   return <Tabs defaultValue={NODE_JS} data-test-id="trigger-code-snippet" menuTabs={prismTabs} />;
 }
 
-export const createNodeSnippet = (
-  identifier: string,
-  to: Record<string, unknown>,
-  payload: Record<string, unknown>,
-  overrides?: Record<string, unknown>,
-  snippet?: Record<string, unknown>,
-  apiKey = '<API_KEY>'
-) => {
-  return `import { Novu } from '@novu/node'; 
-
-const novu = new Novu('${apiKey}');
-
-novu.trigger('${identifier}', ${JSON.stringify(
-    {
-      to,
-      payload,
-      overrides,
-      ...snippet,
-    },
-    null,
-    2
-  )
-    .replace(/"([^"]+)":/g, '$1:')
-    .replace(/"/g, "'")});
-`;
-};
-
 export const getNodeTriggerSnippet = (
   identifier: string,
   to: Record<string, unknown>,
@@ -86,31 +59,6 @@ export const getNodeTriggerSnippet = (
       {triggerCodeSnippet}
     </Prism>
   );
-};
-
-export const createCurlSnippet = (
-  identifier: string,
-  to: Record<string, unknown>,
-  payload: Record<string, unknown>,
-  overrides?: Record<string, unknown>,
-  snippet?: Record<string, unknown>,
-  apiKey = '<REPLACE_WITH_API_KEY>'
-) => {
-  return `curl --location --request POST '${API_ROOT}/v1/events/trigger' \\
---header 'Authorization: ApiKey ${apiKey}' \\
---header 'Content-Type: application/json' \\
---data-raw '${JSON.stringify(
-    {
-      name: identifier,
-      to,
-      payload,
-      overrides,
-      ...snippet,
-    },
-    null,
-    2
-  )}'
-  `;
 };
 
 export const getCurlTriggerSnippet = (
