+++ apps/web/src/studio/components/workflows/test-workflow/WorkflowTestTriggerPanel.tsx
@@ -4,7 +4,13 @@ import { IconOutlineBolt } from '@novu/novui/icons';
 import { Tabs, Title } from '@novu/novui';
 import { HStack } from '@novu/novui/jsx';
 import { FC } from 'react';
-import { createCurlSnippet, createNodeSnippet } from '../../../../pages/templates/components/TriggerSnippetTabs';
+import {
+  createNodeSnippet,
+  createCurlSnippet,
+  createPhpSnippet,
+  createGoSnippet,
+  createPythonSnippet,
+} from '../../../../utils/codeSnippets';
 
 interface IWorkflowTestTriggerPanelProps {
   identifier: string;
@@ -12,111 +18,6 @@ interface IWorkflowTestTriggerPanelProps {
   payload: Record<string, unknown>;
   apiKey: string;
 }
-
-const createPhpSnippet = (
-  identifier: string,
-  to: Record<string, any>,
-  payload: Record<string, any>,
-  apiKey?: string
-) => {
-  return `use Novu\\SDK\\Novu;
-
-$novu = new Novu('${apiKey}');
-
-$response = $novu->triggerEvent([
-    'name' => '${identifier}',
-    'payload' => [
-${Object.keys(payload)
-  .map((key) => {
-    return `        '${key}' => '${payload[key]}',`;
-  })
-  .join('\n')}
-    ],
-    'to' => [
-${Object.keys(to)
-  .map((key) => {
-    return `        '${key}' => '${to[key]}',`;
-  })
-  .join('\n')}
-    ]
-])->toArray();`;
-};
-
-const createPythonSnippet = (
-  identifier: string,
-  to: Record<string, any>,
-  payload: Record<string, any>,
-  apiKey?: string
-) => {
-  return `from novu.api import EventApi
-
-url = "https://api.novu.co"
-
-novu = EventApi(url, "${apiKey}").trigger(
-    name="${identifier}",
-    recipients="${to.subscriberId}",
-    payload={
-${Object.keys(payload)
-  .map((key) => {
-    return `        "${key}":"${payload[key]}",`;
-  })
-  .join('\n')}
-    },
-)`;
-};
-
-const createGoSnippet = (
-  identifier: string,
-  to: Record<string, any>,
-  payload: Record<string, any>,
-  apiKey?: string
-) => {
-  return `package main
-
-import (
-	"context"
-	"fmt"
-	novu "github.com/novuhq/go-novu/lib"
-	"log"
-)
-
-func main() {
-	ctx := context.Background()
-	to := map[string]interface{}{
-${Object.keys(payload)
-  .map((key) => {
-    return `		"${key}": "${payload[key]}",`;
-  })
-  .join('\n')}
-	}
-
-	payload := map[string]interface{}{
-${Object.keys(to)
-  .map((key) => {
-    return `		"${key}": "${to[key]}",`;
-  })
-  .join('\n')}
-	}
-
-	data := novu.ITriggerPayloadOptions{To: to, Payload: payload}
-	novuClient := novu.NewAPIClient("${apiKey}", &novu.Config{})
-
-	resp, err := novuClient.EventApi.Trigger(ctx, "${identifier}", data)
-	if err != nil {
-		log.Fatal("novu error", err.Error())
-		return
-	}
-
-	fmt.Println(resp)
-
-	// get integrations
-	integrations, err := novuClient.IntegrationsApi.GetAll(ctx)
-	if err != nil {
-		log.Fatal("Get all integrations error: ", err.Error())
-	}
-	fmt.Println(integrations)
-}`;
-};
 
 export const WorkflowTestTriggerPanel: FC<IWorkflowTestTriggerPanelProps> = ({ identifier, to, payload, apiKey }) => {
   return (
